+++ tests/Fixtures/Models/Company.php
@@ -5,14 +5,16 @@ declare(strict_types=1);
 namespace RobinsonRyan\HeyYou\Tests\Fixtures\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use RobinsonRyan\HeyYou\Traits\ConfiguresIdentifiers;
 use RobinsonRyan\HeyYou\Traits\Contactable;
 
 /**
- * @property int $id
+ * @property string $id
  * @property string $legal_name
  */
 final class Company extends Model
 {
+    use ConfiguresIdentifiers;
     use Contactable;
 
     protected $table = 'companies';