+++ tests/Fixtures/Models/User.php
@@ -5,15 +5,17 @@ declare(strict_types=1);
 namespace RobinsonRyan\HeyYou\Tests\Fixtures\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use RobinsonRyan\HeyYou\Traits\ConfiguresIdentifiers;
 use RobinsonRyan\HeyYou\Traits\Contactable;
 
 /**
- * @property int $id
+ * @property string $id
  * @property string $name
  * @property string $email
  */
 final class User extends Model
 {
+    use ConfiguresIdentifiers;
     use Contactable;
 
     protected $table = 'users';